src/frontend/requirements.txt
# pyarrow==9.0.0
# referencing==0.32.1
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
bleach==5.0.1
blinker==1.5
cachetools==5.2.0
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
contourpy==1.2.0
cycler==0.12.1
commonmark==0.9.1
cryptography==37.0.4
debugpy==1.6.3
decorator==5.1.1
distlib==0.3.4
docutils==0.19
entrypoints==0.4
executing==0.10.0
filelock==3.6.0
gitdb==4.0.9
fonttools==4.47.2
packaging==21.3
GitPython==3.1.27
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
ipywidgets==8.0.1
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.13.0
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-widgets==3.0.2
keyring==23.8.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
numpy==1.23.2
mdurl==0.1.2
markdown-it-py==3.0.0
kiwisolver==1.4.5
jsonschema-specifications==2023.12.1
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pkginfo==1.8.3
platformdirs==2.5.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
pure-eval==0.2.2
pycparser==2.21
pydeck==0.7.1
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
ipythonipykernel
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==23.2.1
readme-renderer==36.0
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.5.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
seaborn==0.13.1
rpds-py==0.17.1
stack-data==0.4.0
streamlit==1.12.0
streamlit-cookies-manager==0.2.0
streamlit-lottie==0.0.3
streamlit-option-menu==0.3.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.3.0
trycourier==4.4.0
twine==4.0.1
typing_extensions==4.3.0
tenacity==8.2.3
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.11
validators==0.20.0
virtualenv==20.14.1
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==4.0.2
wincertstore==0.2
zipp==3.8.1